+++ coders/msl.c
@@ -529,11 +529,11 @@ MSLPushImage(MSLInfo *msl_info,Image *image)
   msl_info->n++;
   n=msl_info->n;
   MagickReallocMemory(ImageInfo **,msl_info->image_info,
-                      (n+1)*sizeof(ImageInfo *));
+                      ((size_t)n+1)*sizeof(ImageInfo *));
   MagickReallocMemory(DrawInfo **,msl_info->draw_info,
-                      (n+1)*sizeof(DrawInfo *));
-  MagickReallocMemory(Image **,msl_info->attributes,(n+1)*sizeof(Image *));
-  MagickReallocMemory(Image **,msl_info->image,(n+1)*sizeof(Image *));
+                      ((size_t)n+1)*sizeof(DrawInfo *));
+  MagickReallocMemory(Image **,msl_info->attributes,((size_t)n+1)*sizeof(Image *));
+  MagickReallocMemory(Image **,msl_info->image,((size_t)n+1)*sizeof(Image *));
   if ((msl_info->image_info == (ImageInfo **) NULL) ||
       (msl_info->draw_info == (DrawInfo **) NULL) ||
       (msl_info->attributes == (Image **) NULL) ||
@@ -1961,7 +1961,7 @@ MSLStartElement(void *context,const xmlChar *name,
             new_group_info =
               MagickReallocateResourceLimitedClearedArray(MSLGroupInfo *,
                                                           msl_info->group_info,
-                                                          msl_info->nGroups+1,
+                                                          (size_t)msl_info->nGroups+1,
                                                           sizeof(MSLGroupInfo));
             if (new_group_info != (MSLGroupInfo *) NULL)
               {
