+++ magick/colorspace.c
@@ -285,9 +285,9 @@ XYZTransformPackets(void *mutable_data,          /* User provided mutable data *
       y_index = ScaleQuantumToMap(pixels[i].green);
       z_index = ScaleQuantumToMap(pixels[i].blue);
 
-      r = xform->x[x_index].x + xform->y[y_index].x + xform->z[z_index].x + xform->primary_info.x;
-      g = xform->x[x_index].y + xform->y[y_index].y + xform->z[z_index].y + xform->primary_info.y;
-      b = xform->x[x_index].z + xform->y[y_index].z + xform->z[z_index].z + xform->primary_info.z;
+      r = (double) (xform->x[x_index].x + xform->y[y_index].x + xform->z[z_index].x + xform->primary_info.x);
+      g = (double) (xform->x[x_index].y + xform->y[y_index].y + xform->z[z_index].y + xform->primary_info.y);
+      b = (double) (xform->x[x_index].z + xform->y[y_index].z + xform->z[z_index].z + xform->primary_info.z);
 
       r = r < 0.0 ? 0.0 : r > MaxMapDouble ? MaxMapDouble : (r + 0.5);
       g = g < 0.0 ? 0.0 : g > MaxMapDouble ? MaxMapDouble : (g + 0.5);
@@ -1293,9 +1293,9 @@ RGBTransformPackets(void *mutable_data,         /* User provided mutable data */
       g_index = ScaleQuantumToMap(pixels[i].green);
       b_index = ScaleQuantumToMap(pixels[i].blue);
 
-      r = xform->r[r_index].r + xform->g[g_index].r + xform->b[b_index].r;
-      g = xform->r[r_index].g + xform->g[g_index].g + xform->b[b_index].g;
-      b = xform->r[r_index].b + xform->g[g_index].b + xform->b[b_index].b;
+      r = (double) (xform->r[r_index].r + xform->g[g_index].r + xform->b[b_index].r);
+      g = (double) (xform->r[r_index].g + xform->g[g_index].g + xform->b[b_index].g);
+      b = (double) (xform->r[r_index].b + xform->g[g_index].b + xform->b[b_index].b);
 
       r = r < 0.0 ? 0.0 : r > MaxMapDouble ? MaxMapDouble : (r + 0.5);
       g = g < 0.0 ? 0.0 : g > MaxMapDouble ? MaxMapDouble : (g + 0.5);
